# trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Trading Service API",
        "title": "Trading Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/tax": {
            "get": {
                "description": "Vraća listu najskorijih neplaćenih poreskih obaveza (za poslednji obračunati mesec/godinu) za sve korisnike. Za svakog korisnika proverava i da li je registrovan kao aktuar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tax"
                ],
                "summary": "Dohvatanje poslednjeg neplaćenog poreza za sve korisnike",
                "responses": {
                    "200": {
                        "description": "Lista poslednjih neplaćenih poreskih obaveza",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.TaxResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Greška pri izvršavanju upita u bazi (kako je implementirano u kodu)",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greška pri čitanju rezultata iz baze",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/tax/dashboard/{userID}": {
            "get": {
                "description": "Vraća sumu plaćenog poreza za tekuću godinu i sumu neplaćenog poreza za tekući mesec za specificiranog korisnika.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tax"
                ],
                "summary": "Dohvatanje agregiranih poreskih podataka za korisnika",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 123,
                        "description": "ID korisnika čiji se podaci traže",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agregirani poreski podaci za korisnika",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.AggregatedTaxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Neispravan ID korisnika (nije validan broj ili \u003c= 0)",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Interna greška servera pri dohvatanju podataka iz baze",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/tax/run": {
            "post": {
                "description": "Endpoint namenjen za pokretanje procesa obračuna poreza za korisnike. Trenutno nije implementiran i uvek vraća grešku 500.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tax"
                ],
                "summary": "Pokretanje obračuna poreza",
                "responses": {
                    "202": {
                        "description": "Zahtev za obračun poreza je primljen",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greska",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "types.AggregatedTaxResponse": {
            "type": "object",
            "properties": {
                "is_actuary": {
                    "type": "boolean"
                },
                "paid_this_year": {
                    "type": "number"
                },
                "unpaid_this_month": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "types.TaxResponse": {
            "type": "object",
            "properties": {
                "is_actuary": {
                    "type": "boolean"
                },
                "is_paid": {
                    "type": "boolean"
                },
                "tax_amount": {
                    "type": "number"
                },
                "taxable_profit": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Unesite token. Primer: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}